webapp/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "logfire[asgi,fastapi]>=3.15.0",
    "openai>=1.78.0",
    "pydantic-ai-slim[logfire,mcp,openai]>=0.1.11",
    "pydantic-settings>=2.9.1",
    "uvicorn>=0.34.2",
]
